DCS-2079 rename middleware to describe intent
