+++ server/middleware/backTrackPreventionMiddleware.ts
